refactor(plugins): export MessageScopeClassifier from the classifiers package
Callers import it from the package rather than reaching a module deeper, which is the shorter half
of the same path. __all__ is explicit because mypy otherwise reads the import in __init__ as an
implicit re-export and flags it.
