+++ sparkth/plugins/chat/classifiers/__init__.py
@@ -1,0 +1,5 @@
+"""The chat plugin's classifiers, exported so callers import them from the package."""
+
+from sparkth.plugins.chat.classifiers.message_scope import MessageScopeClassifier
+
+__all__ = ["MessageScopeClassifier"]
