+++ sparkth/plugins/chat/routes/completions.py
@@ -19,7 +19,7 @@ from sparkth.lib.llm import (
 )
 from sparkth.lib.log import get_logger
 from sparkth.lib.models import User
-from sparkth.plugins.chat.classifiers.message_scope import MessageScopeClassifier
+from sparkth.plugins.chat.classifiers import MessageScopeClassifier
 from sparkth.plugins.chat.config import ChatSettings, get_chat_settings
 from sparkth.plugins.chat.constants import LLM_PROVIDER_API_ERRORS
 from sparkth.plugins.chat.exceptions import RAGIntentRouterError
